§7 mechanism: causal validation of LoRA promoter heads + paper updates
NEW MECHANISTIC FINDINGS (M3 Pro, this session)
-----------------------------------------------
1. Promoter ablation on the LoRA model (8 heads identified by attention
   routing: L32H3, L32H7, L31H12, L31H15, L32H10, L30H11, L32H0, L32H14).
   Ablating these causes P(v_last) at L32 to drop from 0.98 to 0.70
   (Δ -0.28), but L33+ has redundant pathways that recover to 1.0. The
   promoter heads are causally responsible for v_last building up at L32;
   downstream redundancy absorbs partial ablation.

2. Bootstrap 95% CIs for all probing/logit-lens/attention deltas (R2).
   All deltas exclude zero with tight intervals. PI-correctness L33:
   0.583 [0.437, 0.729] → 0.939 [0.915, 0.963]. P(v_last) L35 PI K=2/N=5:
   0.249 [0.167, 0.332] → 1.000 [1.000, 1.000]. Top attention-routing
   head L32H3: 0.107 [0.088, 0.126] → 0.784 [0.757, 0.809].

3. Attention routing at reversal regime K=15/N=20. Same L30-L33 promoter
   circuit operates in reversal regime — all 8 primacy-regime top heads
   are also top promoters at K=15/N=20 with comparable Δ. Four additional
   heads activate (L27H6, L27H1, L27H5, L28H10, L29H4) likely specific to
   high-K routing. Mechanism is regime-general.

4. *** REFRAMING from the baseline-promoter control run *** Ablating the
   same 8 heads in the BASELINE model (no LoRA) drops L32 P(v_last) by
   Δ -0.38 — actually LARGER than the LoRA-side Δ -0.28. The 8 heads are
   NOT LoRA-installed. They carry partial v_last attention pre-LoRA;
   LoRA AMPLIFIES that signal and installs new L33+ downstream
   redundancy. The downstream-redundancy piece is the genuinely new
   thing: baseline ablation propagates to L35 (Δ -0.09); LoRA ablation
   is fully absorbed by L33+ (Δ 0.00).

These four signals (probing + logit lens + attention routing + Stage 3C
across two head sets) converge on layers L24-L33, peaking L30-L33, with
the reframed claim "amplify + add downstream redundancy."

NEW SCRIPTS IN lora_intervention/
---------------------------------
- run_probing_baseline.py: baseline probing wrapper with bf16 option
  (needed for K=15/N=20 which OOMs on MPS at fp32).
- run_stage3_promoter_ablation.py: ablate LoRA's L30-L33 promoter heads.
- run_stage3_baseline_promoter_ablation.py: control — same 8 heads in
  base model.
- compute_cis.py: bootstrap CIs from existing JSON outputs (no GPU).
- PLAN_baseline_promoter_ablation.md: spec for the control run.

run_probing_lora.py: --dtype flag added (float32 default, bfloat16 for
heavy operating points).

v3/scripts/experiments/probing_classifier.py: one-line fix — `.float()`
before `.numpy()` so bf16 forwards work end-to-end.

PAPER
-----
- paper/main.tex: substantial expansion (1074 lines changed).
- paper/SECTION7_MECHANISM_PLAN.md: section 7 plan document with
  per-subsection data inventory, status, remaining tasks, and reviewer
  critique table. Updated to reflect today's runs.
- paper/PAPER_PLAN.md: session 2026-05-25 progress block.
- Figures regenerated: main results, training dynamics (smollm2 +
  smollm3), dense IVQ panels, multimodel IVQ, gemma 4b CVQ, new mechanism
  figure.
- Tables added: tab_ablation_fourway, tab_mechanism, tab_lora_arb_full,
  tab_format, tab1_proprietary, tab2_openweight.
- Scripts: generate_main_results extended; new measure_context_usage,
  plot_gemma4b_fvq_cvq.

NOT IN THIS COMMIT (per repo size policy)
-----------------------------------------
- 165 MB attention-routing JSONs at K=2/N=30 and K=15/N=20 reversal for
  the LoRA model. The corresponding summary .txt files capture every
  publishable number; the run_attention_routing_lora.py script regenerates
  the raw JSON in 5-20 min as needed.
- 15 MB logit-lens JSON for K=2/N={5,10,50} LoRA run.
- The three tarballs (lora_intervention.tar.gz, v3/results_vllm.tar.gz,
  qwen35_results.tar.gz) — gitignored.

+++ v3/scripts/experiments/probing_classifier.py
@@ -122,7 +122,7 @@ def collect_representations(model, tokenizer, value_pool, model_name,
             # Extract residual stream at answer position (last token) for each layer
             reps = np.zeros((n_layers, d_model))
             for layer in range(n_layers):
-                resid = cache["resid_post", layer][0, -1, :].cpu().numpy()
+                resid = cache["resid_post", layer][0, -1, :].float().cpu().numpy()
                 reps[layer] = resid
 
             # Get model's prediction
